% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19CohortEvaluation.R
\docType{package}
\name{Covid19CohortEvaluation-package}
\alias{Covid19CohortEvaluation}
\alias{Covid19CohortEvaluation-package}
\title{Covid19CohortEvaluation: A Central Package for Evaluation the Cohorts Used in the Various COVID-19 Studies}
\description{
A central package for evaluation the cohorts used in the various COVID-19 studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi-studies.github.io/Covid19CohortEvaluation}
  \item \url{https://github.com/ohdsi-studies/Covid19CohortEvaluation}
  \item Report bugs at \url{https://github.com/ohdsi-studies/Covid19CohortEvaluation/issues}
}

}
\keyword{internal}
