% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{getCohortGroups}
\alias{getCohortGroups}
\title{Get the list of cohort groups}
\usage{
getCohortGroups()
}
\value{
A character vector of cohort group names included in this package.
}
\description{
Get the list of cohort groups
}
